Split into separate crates and binaries
We know we want to have this eventually because the orderbook will run
indendent of the solver. This changes the structure of the code a lot so
it makes sense to do it early when it is less work.
The plan is that solver communicates with the orderbook through an http
api.
In the orderbook crate we continue to have the `struct Orderbook` that
manages the orderbook while in the solver crate we will probably have a
`trait Orderbook` that allows reading the orderbook.
