+++ Cargo.toml
@@ -1,16 +1,5 @@
-[package]
-name = "oba-services"
-version = "0.1.0"
-authors = ["josojo <josojo@hotmail.de>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.33"
-ethcontract = { version = "0.8.0", default-features = false }
-hex-literal = "0.3.1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.59"
-tokio = { version = "0.2.4", features =[ "macros"] }
-warp = "0.2"
+[workspace]
+members = [
+    "orderbook",
+    "solver",
+]
