+++ orderbook/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "orderbook"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0.33"
+ethcontract = { version = "0.8.0", default-features = false }
+hex-literal = "0.3.1"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.59"
+tokio = { version = "0.2.4", features =[ "macros"] }
+warp = "0.2"
