+++ solver/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "solver"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0.33"
+ethcontract = { version = "0.8.0", default-features = false }
+tokio = { version = "0.2.4", features =[ "macros"] }
