+++ solver/src/main.rs
@@ -1,0 +1,6 @@
+mod batcher;
+
+#[tokio::main]
+async fn main() {
+    todo!("run driver")
+}
