--- src/batcher/solve_pair.rs
@@ -1,10 +1,0 @@
-use crate::models::Order;
-use crate::models::Solution;
-use anyhow::Result;
-
-pub fn solve_pair(
-    _sell_orders_token0: Vec<Order>,
-    _sell_orders_token1: Vec<Order>,
-) -> Result<Solution> {
-    unimplemented!();
-}
