--- src/batcher/submit_solution.rs
@@ -1,6 +1,0 @@
-use crate::models::solution::Solution;
-use anyhow::Result;
-
-pub async fn submit_solution(_solution: Solution) -> Result<()> {
-    unimplemented!();
-}
